requirements.txt
# Core dependencies
Django==3.2.16
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.2
django-environ==0.9.0
psycopg2-binary==2.9.5
gunicorn==20.1.0

# Authentication
djangorestframework-simplejwt==5.2.2
django-allauth==0.51.0


# PDF processing
PyPDF2==2.11.2
reportlab==3.6.12

# Utilities
python-dateutil==2.8.2
requests==2.28.1
Pillow==9.3.0
pytz==2022.6

